Refactor forge/neoforge plugins
