+++ frontend-dsl/forge/src/main/groovy/org/groovymc/modsdotgroovy/frontend/forge/DependenciesBuilder.groovy
@@ -1,4 +1,4 @@
-package org.groovymc.modsdotgroovy.frontend
+package org.groovymc.modsdotgroovy.frontend.forge
 
 import groovy.transform.CompileDynamic
 import groovy.transform.CompileStatic
@@ -7,6 +7,8 @@ import groovy.transform.stc.SimpleType
 import groovy.util.logging.Log4j2
 import org.groovymc.modsdotgroovy.core.ModsDotGroovyCore
 import org.groovymc.modsdotgroovy.core.versioning.VersionRangeAware
+import org.groovymc.modsdotgroovy.frontend.DslBuilder
+import org.groovymc.modsdotgroovy.frontend.MapClosureInterceptor
 import org.jetbrains.annotations.Nullable
 
 import static groovy.lang.Closure.DELEGATE_FIRST
@@ -21,7 +23,7 @@ class DependenciesBuilder extends DslBuilder implements MapClosureInterceptor {
     @Nullable String minecraft = null
 
     void mod(@DelegatesTo(value = DependencyBuilder, strategy = DELEGATE_FIRST)
-             @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.DependencyBuilder')
+             @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.forge.DependencyBuilder')
              final Closure closure) {
         log.debug "mod(closure)"
         core.push('dependency')
@@ -34,7 +36,7 @@ class DependenciesBuilder extends DslBuilder implements MapClosureInterceptor {
 
     void mod(final String modId,
              @DelegatesTo(value = DependencyBuilder, strategy = DELEGATE_FIRST)
-             @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.DependencyBuilder')
+             @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.forge.DependencyBuilder')
              final Closure closure) {
         log.debug "mod(string, closure)"
         core.push('dependency')
