+++ frontend-dsl/forge/src/main/groovy/org/groovymc/modsdotgroovy/frontend/forge/FeaturesBuilder.groovy
@@ -1,8 +1,10 @@
-package org.groovymc.modsdotgroovy.frontend
+package org.groovymc.modsdotgroovy.frontend.forge
 
 import groovy.transform.CompileStatic
 import groovy.util.logging.Log4j2
 import org.groovymc.modsdotgroovy.core.ModsDotGroovyCore
+import org.groovymc.modsdotgroovy.frontend.DslBuilder
+import org.groovymc.modsdotgroovy.frontend.PropertyInterceptor
 import org.jetbrains.annotations.Nullable
 
 @CompileStatic
