+++ frontend-dsl/forge/src/main/groovy/org/groovymc/modsdotgroovy/frontend/forge/ModInfoBuilder.groovy
@@ -1,9 +1,12 @@
-package org.groovymc.modsdotgroovy.frontend
+package org.groovymc.modsdotgroovy.frontend.forge
 
 import groovy.transform.CompileStatic
 import groovy.transform.stc.ClosureParams
 import groovy.transform.stc.SimpleType
 import groovy.util.logging.Log4j2
+import org.groovymc.modsdotgroovy.frontend.DslBuilder
+import org.groovymc.modsdotgroovy.frontend.MapClosureInterceptor
+import org.groovymc.modsdotgroovy.frontend.PropertyInterceptor
 import org.jetbrains.annotations.Nullable
 import org.groovymc.modsdotgroovy.core.ModsDotGroovyCore
 
@@ -75,7 +78,7 @@ class ModInfoBuilder extends DslBuilder implements PropertyInterceptor, MapClosu
     String author = ''
 
     void dependencies(@DelegatesTo(value = DependenciesBuilder, strategy = DELEGATE_FIRST)
-                      @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.DependenciesBuilder')
+                      @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.forge.DependenciesBuilder')
                       final Closure closure) {
         log.debug "dependencies(closure)"
         core.push('dependencies')
@@ -87,7 +90,7 @@ class ModInfoBuilder extends DslBuilder implements PropertyInterceptor, MapClosu
     }
 
     void features(@DelegatesTo(value = FeaturesBuilder, strategy = DELEGATE_FIRST)
-                  @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.FeaturesBuilder')
+                  @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.forge.FeaturesBuilder')
                   final Closure closure) {
         log.debug "features(closure)"
         core.push('features')
