+++ frontend-dsl/neoforge/src/main/groovy/DependencyOrdering.groovy
@@ -1,0 +1,8 @@
+import groovy.transform.CompileStatic
+
+@CompileStatic
+enum DependencyOrdering {
+    BEFORE, AFTER, NONE
+
+    DependencyOrdering() {}
+}
