+++ frontend-dsl/neoforge/src/main/groovy/DependencySide.groovy
@@ -1,0 +1,8 @@
+import groovy.transform.CompileStatic
+
+@CompileStatic
+enum DependencySide {
+    CLIENT, SERVER, BOTH
+
+    DependencySide() {}
+}
