+++ frontend-dsl/neoforge/src/main/groovy/DependencyType.groovy
@@ -1,0 +1,8 @@
+
+
+enum DependencyType {
+    REQUIRED,
+    OPTIONAL,
+    INCOMPATIBLE,
+    DISCOURAGED
+}
