+++ frontend-dsl/neoforge/src/main/groovy/org/groovymc/modsdotgroovy/frontend/neoforge/ModsBuilder.groovy
@@ -12,14 +12,14 @@ import org.groovymc.modsdotgroovy.frontend.PropertyInterceptor
 import static groovy.lang.Closure.DELEGATE_FIRST
 
 @CompileStatic
-@Log4j2(category = 'MDG - NeoForge Frontend')
-class NeoForgeModsBuilder extends DslBuilder implements PropertyInterceptor, MapClosureInterceptor {
-    void modInfo(@DelegatesTo(value = NeoForgeModInfoBuilder, strategy = DELEGATE_FIRST)
-                 @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.neoforge.NeoForgeModInfoBuilder')
+@Log4j2(category = 'MDG - Forge Frontend')
+class ModsBuilder extends DslBuilder implements PropertyInterceptor, MapClosureInterceptor {
+    void modInfo(@DelegatesTo(value = ModInfoBuilder, strategy = DELEGATE_FIRST)
+                 @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.neoforge.ModInfoBuilder')
                  final Closure closure) {
-        log.debug 'modInfo(closure)'
+        log.debug "modInfo(closure)"
         core.push('modInfo')
-        final modInfoBuilder = new NeoForgeModInfoBuilder(core)
+        final modInfoBuilder = new ModInfoBuilder(core)
         closure.resolveStrategy = DELEGATE_FIRST
         closure.delegate = modInfoBuilder
         closure.call(modInfoBuilder)
@@ -27,12 +27,11 @@ class NeoForgeModsBuilder extends DslBuilder implements PropertyInterceptor, Map
     }
 
     void modInfo(final String modId,
-                 @DelegatesTo(value = NeoForgeModInfoBuilder, strategy = DELEGATE_FIRST)
-                 @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.neoforge.NeoForgeModInfoBuilder')
-                 final Closure closure) {
-        log.debug 'modInfo(closure)'
+                 @DelegatesTo(value = ModInfoBuilder, strategy = DELEGATE_FIRST)
+                 @ClosureParams(value = SimpleType, options = 'org.groovymc.modsdotgroovy.frontend.neoforge.ModInfoBuilder') final Closure closure) {
+        log.debug "modInfo(closure)"
         core.push('modInfo')
-        final modInfoBuilder = new NeoForgeModInfoBuilder(core)
+        final modInfoBuilder = new ModInfoBuilder(core)
         core.put('modId', modId)
         closure.resolveStrategy = DELEGATE_FIRST
         closure.delegate = modInfoBuilder
@@ -40,7 +39,7 @@ class NeoForgeModsBuilder extends DslBuilder implements PropertyInterceptor, Map
         core.pop()
     }
 
-    NeoForgeModsBuilder(final ModsDotGroovyCore core) {
+    ModsBuilder(final ModsDotGroovyCore core) {
         super(core)
     }
 }
