+++ stock-plugins/forgelike/src/main/resources/META-INF/services/org.groovymc.modsdotgroovy.plugin.ModsDotGroovyPlugin
@@ -1,0 +1,1 @@
+org.groovymc.modsdotgroovy.plugin.ForgeLikePlugin
