+++ stock-plugins/multiplatform/src/main/resources/META-INF/services/org.groovymc.modsdotgroovy.plugin.ModsDotGroovyPlugin
@@ -1,2 +1,1 @@
 org.groovymc.modsdotgroovy.plugin.MultiplatformPlugin
-org.groovymc.modsdotgroovy.plugin.MultiplatformFabricNormalizationPlugin
